feat(package): eliminate the need for compilation of dependencies on most platforms
This commit bumps the `sharp` depenecy to v0.20.0, which introduces prebuilt binaries for `sharp.node`. Sharp was the only remaining dependency which did not provide pre-built binaries, and therefore always required compilation.

Now that all deps have pre-built binaries, this library will only need to have a compiler toolchain on platforms which are not supported by the prebuilt binaries.